Move all datastore logic to useDatastore
